Distributions (#1)
* distribution: beta: Initial port.

* specs: Load all classes and libs when running rspec.

* distributions: Beta distribution.

This change adds the Beta distribution with the following methods:
  * Beta function. If x,y not specified, it will use alpha, beta.
  * Probability density function (density_function).
  * Beta mode.
  * Beta mean.

* distributions: Normal and Standard normal distributions.

This change adds two distributions: Normal and Standard Normal.
The following methods are available:
* Probability density function (density_function).
* Normal mode.
* Normal variance.
* Standard Normal mode.
* Standard Normal variance.

* distributions: beta: Ensure float values in alpha, beta.

* distributions: Chi-squared distribution.

This change includes the following methods for the Chi-squared distribution:

* Probability density function.
* Mode.
* Variance.

* distributions: beta: Extract beta function to class method.

* distributions: F distribution.

This change adds the following methods for the F distribution:
* Probability density function.
* Mean.
* Mode.

* distributions: T-Student distribution.

This change adds the following methods for the T-Student distribution:
* Probability density function.
* Mean.
* Variance.
* Mode.

* distributions: Uniform distribution.

This changes adds the following methods for the Uniform distribution:
* Probability density function.
* Mean.
* Median.
* Variance.

* distributions: Add alias to Statistics::Distribution.

This change allows to do Distribution:: instead of Statistics::Distribution::

* fix-me-up-later: Weibull distribution.

* distribution: Weibull distribution.

This change adds the following methods for the Weibull distribution:
* Probability density function.
* Mean.
* Mode.
* Variance.

* Distributions: Cumulative Functions feature. (#2)

* distribution: weibull: Add Cumulative function.

This change allows to calculate probabilities based on the specified
scale and shape at initialize time.

* distribution: uniform: Add cumulative function.

This change allows to calculate the probability using the uniform
distribution for the specified value.

* distribution: weibull: Fix variance formula.

* distribution: beta: Add Cumulative distribution function.

This is a big change. It is composed by two changes:
  1. Implementation of the generalized incomplete beta function.
  2. The introduction of the cumulative distribution function, which
     is the same as the generalized incomplete beta function for the
     beta distribution.

Woot!

* distributions: t-student: Add Cumulative function.

This change allows to calculate probabilites based on the specified
degrees of freedom at initialize time.

* distribution: normal: Add cumulative function.

This change allows to use the CDF for the normal distribution and
the standard normal distribution as well.

* distribution: f: Add Cumulative function.

This change implements the cumulative distribution function for
the F-Distribution.

* distribution: chi-squared: Add cumulative function.

This change also includes a couple of special functions:
  * The composite simpson's rule.
  * The lower incomplete gamma function.

* math: Override math module to include special functions.

* statistics: Implement special functions in Math mixin.

* math: Implement some basic statistical functions.

* Factorial for integer values.
* Combinations.
* Permutations.

* distributions: Binomial distribution.

This is the first of multiple discrete distributions. It brings:
* Binomial Cumulative distribution function.
* Binomial Probability mass function.
* Binomial mean.
* Binomial variance.
* Binomial mode.

* squash: spec: Compare real part of result instead of rounded values.

* statistics: Require 'descriptive_statistics' gem.

* distribution: Add Poisson distribution.

Added poisson distribution with probability mass function and
cumulative function. It also includes mean and variance.