t-test: Raise error when standard deviation is zero.
Fixes: #20

This change allow us to know when it's necessary to revisit the
specified samples when we have an standard devation of zero, which
in practice seems to be an edge case. We are implementing this
approach following the answer made to this question
(https://stats.stackexchange.com/questions/78570/t-test-with-sample-standard-deviation-of-zero-possible)
where it's important to reconsider samples with std of zero.