all: Coerce to rational instead of float.
fixes #34

This change allow us to perform operations with BigDecimal in a more
perfomant way, also it sightly improves the accuracy of the calculations
made.

This change also includes a full copy of the current spec suite but
using BigDecimal instead of Numeric or Float, to make sure everything is
working as expected.