Aliasing Distribution Causes Namespace Collision (#46)
* Aliasing Distribution Causes Namespace Collision

**Description**
The gem loads prior to the main application code and aliasing the gems
namespace to a root namespace creates conflicts with main app. This
update removes the aliasing to prevent this from happening.

**Relates to Issue**
https://github.com/estebanz01/ruby-statistics/issues/45

* Version 3.0
removes distribution namespace aliasing

**Description**
This update remove the `Distribution` name space aliasing. To avoid
issues it is recommended that any references to `Distribution` are
replaced with the full name path `Statistics::Distribution` to avoid any
disruption.

Co-authored-by: Rebecca Chapin <msneverheart@gmail.com>
Co-authored-by: Rebecca Chapin <rchapin@Rebeccas-MacBook-Pro.local>