fix: Implement zero division bugfix in lower gamma function calculation.
This change implements a fix, discovered in #78 , where the
incomplete gamma function raises a `ZeroDivisionError` when
X is too small.

This was caused by the liberal use of the `round` function when
calculating the iterations needed to perform the simpson's rule.

Now, with this change, if the X is too small, it rounds it to
the first decimal, then rounds it again after multiplying it by
10_000. Now, if the second round still give us zero, then we default
to 100_000 iterations instead, as we must assume it's a pretty small
value that needs to be accounted for.