Bump to version 3.0.1.
This release fixes an issue with division by zero when coercing to
rationals and fixes another issue where it segfaults when calculating
BigDecimals with big decimal parts. Altough it's fixed on latest
available version, it might generate incompatibilities with older ruby
versions, so it's better to coerce to rational when possible.