Updates on Chi squared and Lower incomplete gamma function (#122)
* feat: Calculate CDF for Xi square using a special case when df = 2.

This change introduces the calculation of the CDF for the xi square
distribution by executing the special case when the degrees of
freedom is 2. This is to make it on par with calculations made in R
and some python packages. It derived from work made by @oliver-czulo
in #115.

* fix: Always use at least 10_000 iterations when calculating lower gamma.

This change fixes a small bug where we were calculating the
lower incomplete gamma function via simpson's rule with lower
than 10_000 iterations, which is not ideal. This was found while
doing some debugging on #115.