Name lib/ top-level file to match expected require
In 00312b9a329c742c2b851a182d3ab35bb86fe535 the namespace of this gem
was converted from `statistics` to `ruby-statistics`. This includes the
instructions for requiring the entire gem to use
`require 'ruby-statistics'`. However, that file does not exist in the
gem.

The `lib/statistics.rb` file still existed. And in tests for the gem,
that file was still being loaded. As a result the tests were still
passing after the change.

To be in agreement with the documentation and the upgrade path, this
creates a `lib/ruby-statistics.rb` file. That is loaded in tests to
confirm the behavior we are documenting in the README.

This file naming is a bit unconventional, as we may expect to be using
underscore to separate the words. That would also be consistent with
RubyGem's [naming guide](https://guides.rubygems.org/name-your-gem/).
However, the gem name is already defined and considered out of scope for
this change.

We *can* make this file `lib/ruby_statistics.rb`. That would match a
Ruby conventional file-naming scheme. As a result of that, we would
update the documentation to `require 'ruby-statistics'`. I would also
advocate for changing the `ruby-statistics` directory to
`ruby_statistics`.

Doing that would have inconsistency between the gem name and what's
being required, though some of it would be more conventional.

Rather than making those sweeping changes, this proposes creating this
file so that the gem can be required as the README suggests.